requirements.txt
alembic
anthropic
click
colorama
et-xmlfile
Flask
Flask-Login
Flask-WTF
google-generativeai
greenlet
itsdangerous
Jinja2
lark
Mako
MarkupSafe
networkx
numpy
openai
openpyxl
pandas
platformdirs
plotly
python-dateutil
pytz
six
sentence-transformers
SQLAlchemy
tenacity
tzdata
Werkzeug
WTForms
zipp